Fix I_GetTime/I_SleepToTic freezing on 32-bit non-Windows

+++ src/sdl/i_system.c
@@ -2224,7 +2224,7 @@ tic_t I_GetTime (void)
 		ts.tv_sec--;
 		ts.tv_nsec += 1000000000;
 	}
-	ticks = ts.tv_sec * 1000000000 + ts.tv_nsec;
+	ticks = ((uint64_t)ts.tv_sec * 1000000000) + ts.tv_nsec;
 	ticks = ticks * TICRATE / 1000000000;
 
 	return (tic_t)ticks;
@@ -2244,8 +2244,8 @@ void I_SleepToTic(tic_t tic)
 		ts.tv_sec--;
 		ts.tv_nsec += 1000000000;
 	}
-	curtime = ts.tv_sec * 1000000000 + ts.tv_nsec;
-	targettime = (uint64_t)tic * 1000000000 / TICRATE;
+	curtime = ((uint64_t)ts.tv_sec * 1000000000) + ts.tv_nsec;
+	targettime = ((uint64_t)tic * 1000000000) / TICRATE;
 	if (targettime < curtime)
 		return;
